adds the --json-file argument
This is a path to a json file that the bz information
will be stored in. If it exists already and is valid json
then bz information will be appeneded with the project
name as the key. This is to allow us to run this script
against multiple project and get a combined list of bzs
from all projects.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>